Fix preset 12

+++ src/pyphonic/12_syncdnoise.py
@@ -1,3 +1,6 @@
+# beat_synced_noise
+# Add some fizz to your chorus with noise. 1000 ticks per beat, so starting at 250 is a 16th after each beat.
+
 import numpy as np
 
 import pyphonic
@@ -10,5 +13,5 @@ def process_npy(midi, audio):
     if start_ticks <= pyphonic.getTransport()["ticks"] < end_ticks:
         noise = np.random.normal(scale=NOISE_VOLUME, size=audio.shape)
         noise = (np.abs(np.random.normal(size=audio.shape)) < NOISE_DENSITY) * noise
-        return midi, audio + noise
+        return midi, audio + noise.astype(np.float32)
     return midi, audio
